better script name and lots of comments
